+++ d2lvalence_util/__init__.py
@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
 
 """
-:module: d2lvalence
-:synopsis: Package for D2L Valence client applications.
+:module: d2lvalence_util
+:synopsis: Support package for D2L Valence client applications.
 :license: Apache v2.0, see LICENSE for more details.
 
 :see: http://www.desire2learn.com/r/valencehome for more information about the Valence project.
 """
 
-__title__ = 'd2lvalence'
-__version__ = '0.1.15'
-__build__ = 0x00010D
+__title__ = 'd2lvalence_util'
+__version__ = '0.1.16'
+__build__ = 0x00010F
 __author__ = 'Desire2Learn Extensibility'
-__license__ = 'Apache v2'
+__license__ = 'Apache 2.0'
 __copyright__ = 'Copyright 2013 Desire2Learn Incorporated.'